feat: добавлен обработчик для propertyValue DKCKZ-346 DKCKZ-420
 - добавлена реализация сервиса действий для propretyValue

 - добавлен DTO, mapper, denormalizer для propertyValue

https://track.ratio.digital/youtrack/issue/DKCKZ-420
