+++ backend/src/Features/Properties/PropertyUnit/DTO/Message/PropertyUnitMessageDTO.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Features\Properties\PropertyUnit\DTO\Message;
+
+use App\Helper\Interface\Message\MessageDTOInterface;
+
+final readonly class PropertyUnitMessageDTO implements MessageDTOInterface
+{
+
+}
