+++ backend/src/Features/Properties/PropertyUnit/Service/PropertyUnitActionService.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Features\Properties\PropertyUnit\Service;
+
+use App\Helper\Interface\ActionInterface;
+use App\Helper\Interface\Message\MessageDTOInterface;
+
+final readonly class PropertyUnitActionService implements ActionInterface
+{
+
+    public function create(MessageDTOInterface $dto): void
+    {
+
+    }
+
+    public function update(MessageDTOInterface $dto): void
+    {
+
+    }
+
+    public function delete(MessageDTOInterface $dto): void
+    {
+
+    }
+}
