+++ backend/src/Features/Properties/PropertyValue/Repository/PropertyValueRepository.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Features\Properties\PropertyValue\Repository;
+
+use App\Document\Properties\PropertyValue;
+use Doctrine\Bundle\MongoDBBundle\{ManagerRegistry, Repository\ServiceDocumentRepository};
+
+class PropertyValueRepository extends ServiceDocumentRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, PropertyValue::class);
+    }
+}
