--- backend/src/Features/PropertyUnit/DTO/Message/PropertyUnitMessageDTO.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Features\PropertyUnit\DTO\Message;
-
-final readonly class PropertyUnitMessageDTO
-{
-
-}
