--- backend/src/Features/PropertyUnit/Service/PropertyUnitActionService.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace App\Features\PropertyUnit\Service;
-
-use App\Helper\Interface\ActionInterface;
-use App\Helper\Interface\Message\MessageDTOInterface;
-
-final readonly class PropertyUnitActionService implements ActionInterface
-{
-
-    public function create(MessageDTOInterface $dto): void
-    {
-
-    }
-
-    public function update(MessageDTOInterface $dto): void
-    {
-
-    }
-
-    public function delete(MessageDTOInterface $dto): void
-    {
-
-    }
-}
