--- backend/src/Features/PropertyValue/DTO/Message/PropertyValueMessageDTO.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Features\PropertyValue\DTO\Message;
-
-use App\Helper\Interface\Message\MessageDTOInterface;
-
-class PropertyValueMessageDTO implements MessageDTOInterface
-{
-
-}
