--- backend/src/Features/PropertyValue/Repository/PropertyValueRepository.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace App\Features\PropertyValue\Repository;
-
-use App\Document\Properties\PropertyValue;
-use Doctrine\Bundle\MongoDBBundle\{ManagerRegistry, Repository\ServiceDocumentRepository};
-
-class PropertyValueRepository extends ServiceDocumentRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, PropertyValue::class);
-    }
-}
