--- backend/src/Features/PropertyValue/Service/PropertyValueActionService.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace App\Features\PropertyValue\Service;
-
-use App\Helper\Interface\ActionInterface;
-use App\Helper\Interface\Message\MessageDTOInterface;
-
-class PropertyValueActionService implements ActionInterface
-{
-
-    public function create(MessageDTOInterface $dto): void
-    {
-
-    }
-
-    public function update(MessageDTOInterface $dto): void
-    {
-
-    }
-
-    public function delete(MessageDTOInterface $dto): void
-    {
-
-    }
-}
